Add --diff-to-file option
Add --diff-to-file option for following tk commands:
- diff
- apply
- prune
- delete

Diff is written to stdout/pager as usual. But it is also written
to file provided via --diff-to-file option. File is truncated when
diff returns no differences.
When we encounter an error we do not guarantee that we
update the file. When we fail to write the diff to file we do not
always abort the workflow (we usually do but tk apply is an
exception - see explanation below).

Notably, when user runs tk apply and the diffing itself fails
then tanka still allows the user to apply the changes. In such
case, the diff file might not be updated (depending on the
nature of the diffing error). This sounds dumb at first but it
actually makes sense. Consider situation when you need to
hotfix something but failing diff prevents you from applying
changes. Theoretically, diffs should never fail for applyable
changes. Practically, diffing performs kubectl dry run which
is inherently more fragile than the apply. Imho this justifies
the design decision to allow apply for changes that fail to diff.

Upstream issue: grafana#588